Obstaculos finalizados

+++ include/Entidades/Personagens/OVNI.h
@@ -7,9 +7,9 @@ namespace Entidades{
     namespace Personagens{
         class OVNI: public Inimigo{
             private:
-                int coef_vel;
+                float coef_vel;
             public:
-                OVNI(Vector2f POS=Vector2f(0.f, 0.f), int COEF_VEL=0);
+                OVNI(Vector2f POS=Vector2f(0.f, 0.f), float COEF_VEL=0);
                 ~OVNI();
                 void perseguir(Vector2f dist);
                 void mover_se();
